requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiofile==3.9.0
    # via py-key-value-aio
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   mcp
    #   py-key-value-aio
    #   sse-starlette
    #   starlette
    #   watchfiles
attrs==25.4.0
    # via
    #   cyclopts
    #   jsonschema
    #   referencing
authlib==1.6.9
    # via fastmcp
backports-tarfile==1.2.0
    # via jaraco-context
beartype==0.22.9
    # via py-key-value-aio
cachetools==7.0.2
    # via py-key-value-aio
caio==0.9.25
    # via aiofile
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
cffi==2.0.0
    # via cryptography
click==8.3.1
    # via uvicorn
cryptography==46.0.5
    # via
    #   authlib
    #   pyjwt
cyclopts==4.7.0
    # via fastmcp
dnspython==2.8.0
    # via email-validator
docstring-parser==0.17.0
    # via cyclopts
docutils==0.22.4
    # via rich-rst
email-validator==2.3.0
    # via pydantic
exceptiongroup==1.3.1
    # via fastmcp
fastmcp==3.1.0
    # via edma-mcp (pyproject.toml)
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   edma-mcp (pyproject.toml)
    #   fastmcp
    #   mcp
httpx-sse==0.4.3
    # via mcp
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
importlib-metadata==8.7.1
    # via
    #   keyring
    #   opentelemetry-api
jaraco-classes==3.4.0
    # via keyring
jaraco-context==6.1.0
    # via keyring
jaraco-functools==4.4.0
    # via keyring
jsonref==1.1.0
    # via fastmcp
jsonschema==4.26.0
    # via mcp
jsonschema-path==0.4.5
    # via fastmcp
jsonschema-specifications==2025.9.1
    # via jsonschema
keyring==25.7.0
    # via py-key-value-aio
markdown-it-py==4.0.0
    # via rich
mcp==1.26.0
    # via fastmcp
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.8.0
    # via
    #   jaraco-classes
    #   jaraco-functools
openapi-pydantic==0.5.1
    # via fastmcp
opentelemetry-api==1.40.0
    # via fastmcp
packaging==26.0
    # via fastmcp
pathable==0.5.0
    # via jsonschema-path
platformdirs==4.9.2
    # via fastmcp
py-key-value-aio==0.4.4
    # via fastmcp
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   edma-mcp (pyproject.toml)
    #   fastmcp
    #   mcp
    #   openapi-pydantic
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via mcp
pygments==2.19.2
    # via rich
pyjwt==2.11.0
    # via mcp
pyperclip==1.11.0
    # via fastmcp
python-dotenv==1.2.2
    # via
    #   fastmcp
    #   pydantic-settings
python-multipart==0.0.22
    # via mcp
pyyaml==6.0.3
    # via
    #   fastmcp
    #   jsonschema-path
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
rich==14.3.3
    # via
    #   cyclopts
    #   fastmcp
    #   rich-rst
rich-rst==1.3.2
    # via cyclopts
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
sse-starlette==3.3.2
    # via mcp
starlette==0.52.1
    # via
    #   edma-mcp (pyproject.toml)
    #   mcp
    #   sse-starlette
typing-extensions==4.15.0
    # via
    #   anyio
    #   exceptiongroup
    #   mcp
    #   opentelemetry-api
    #   py-key-value-aio
    #   pydantic
    #   pydantic-core
    #   referencing
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   mcp
    #   pydantic
    #   pydantic-settings
uncalled-for==0.2.0
    # via fastmcp
uvicorn==0.41.0
    # via
    #   edma-mcp (pyproject.toml)
    #   fastmcp
    #   mcp
watchfiles==1.1.1
    # via fastmcp
websockets==16.0
    # via fastmcp
zipp==3.23.0
    # via importlib-metadata
networkx
